Bones of new route table
