+++ source/ed/notifications/module.h
@@ -1,0 +1,15 @@
+#ifndef _ED_SLOT_MODULE_H_
+#define _ED_SLOT_MODULE_H_
+
+
+namespace ed
+{
+  namespace slot
+  {
+    class module
+    {
+    }
+  };
+};
+
+#endif
