+++ source/ed/slot/event.h
@@ -1,0 +1,19 @@
+#ifndef _ED_SLOT_EVENT_H_
+#define _ED_SLOT_EVENT_H_
+
+#include "../
+#include "../notifications/listener.h"
+#include <vector>
+
+namespace ed
+{
+  namespace slot
+  {
+    struct event
+    {
+      std::vector<slot_data::listener> data;
+    };
+  };
+};
+
+#endif
