+++ source/ed/slot/instance.h
@@ -1,0 +1,17 @@
+#ifndef _ED_SLOT_INSTANCE_H_
+#define _ED_SLOT_INSTANCE_H_
+
+#include "module.h"
+
+namespace ed
+{
+  namespace slot
+  {
+    class instance
+    {
+      std::vector<module> data;
+    };
+  };
+};
+
+#endif
