+++ source/ed/slot/route.h
@@ -1,0 +1,20 @@
+#ifndef _ED_SLOT_ROUTE_H_
+#define _ED_SLOT_ROUTE_H_
+
+#include "instance.h"
+
+namespace ed
+{
+  namespace slot
+  {
+    class route
+    {
+      std::vector<instance> data;
+    public:
+
+
+    };
+  };
+};
+
+#endif
